pyproject.toml
[tool.poetry]
name = "foris"
version = "0.1.0"
description = ""
authors = ["Oscar Giraldo Castillo <oscar.gi.cast@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pyhamcrest = "^2.0.4"
six = "^1.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
